+++ CHANGELOG.md
@@ -4,6 +4,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Configs to control number of gem types per biome in the overworld [#448]
+- Config options for geode generation chances [#448]
+### Changed
+- Setting world gen region sizes to zero will now disable gem generation in that dimension [#448]
 ### Fixed
 - Crash when teleporting on servers [#447]